Misc UI updates, validations, project structure tweaks
