+++ Feedbridge/Models/Baby.swift
@@ -14,8 +14,8 @@
 @preconcurrency import FirebaseFirestore
 import Foundation
 
-/// Represents a baby and their associated health tracking data
 // periphery:ignore
+/// Represents a baby and their associated health tracking data
 struct Baby: Identifiable, Codable, Sendable {
     /// Unique identifier for the baby
     @DocumentID var id: String?
