+++ Feedbridge/Models/FeedEntry.swift
@@ -25,8 +25,8 @@ enum MilkType: String, Codable {
     case formula
 }
 
-/// Stores feeding-related data
 // periphery:ignore
+/// Stores feeding-related data
 struct FeedEntry: Identifiable, Codable, Sendable {
     /// Use UUID to generate a unique identifier for Firebase
     @DocumentID var id: String?
