+++ Feedbridge/Models/WeightEntry.swift
@@ -11,8 +11,8 @@
 @preconcurrency import FirebaseFirestore
 import Foundation
 
-/// Stores weight measurements (accepts grams, kilograms, or pounds and ounces)
 // periphery:ignore
+/// Stores weight measurements (accepts grams, kilograms, or pounds and ounces)
 struct WeightEntry: Identifiable, Codable, Sendable {
     @DocumentID var id: String?
