+++ Feedbridge/Views/AddBabyView.swift
@@ -20,11 +20,17 @@ struct AddBabyView: View {
     @State private var babies: [(id: Int, baby: Baby)] = [(id: 0, baby: Baby(name: "", dateOfBirth: Date()))]
     @State private var showAlert = false
     @State private var errorMessage = ""
+    @State private var existingBabies: [Baby] = []
+    @State private var isLoading = true
     
     var body: some View {
         OnboardingView(
             contentView: {
-                VStack(spacing: 24) {
+                Group {
+                    if isLoading {
+                        ProgressView()
+                    } else {
+                        VStack(spacing: 24) {
                     OnboardingTitleView(
                         title: "Add Your Baby",
                         subtitle: "Please enter your baby's information"
@@ -41,6 +47,12 @@ struct AddBabyView: View {
                                 }
                             ))
                             .textFieldStyle(.roundedBorder)
+                            
+                            if !baby.baby.name.isEmpty && isDuplicateName(baby.baby.name, forBabyId: baby.id) {
+                                Text("This name is already taken")
+                                    .foregroundColor(.red)
+                                    .font(.caption)
+                            }
                             
                             DatePicker(
                                 "Date of Birth",
@@ -70,17 +82,28 @@ struct AddBabyView: View {
                     }
                     .padding(.vertical)
                 }
+                    }
+                }
+                .padding()
             },
             actionView: {
-                OnboardingActionsView(
-                    "Continue",
-                    action: {
-                        Task {
-                            await saveBabies()
+                VStack {
+                    OnboardingActionsView(
+                        "Continue",
+                        action: {
+                            Task {
+                                await saveBabies()
+                            }
                         }
+                    )
+                    .disabled(babies.contains(where: { $0.baby.name.isEmpty }) || hasDuplicateNames || isLoading)
+                    
+                    Button("Add babies later") {
+                        onboardingNavigationPath.nextStep()
                     }
-                )
-                .disabled(babies.contains(where: { $0.baby.name.isEmpty }))
+                    .buttonStyle(.automatic)
+                    .padding(.top, 8)
+                }
             }
         )
         .alert("Error", isPresented: $showAlert) {
@@ -88,9 +111,51 @@ struct AddBabyView: View {
         } message: {
             Text(errorMessage)
         }
+        .task {
+            await loadExistingBabies()
+        }
+    }
+    
+    private var hasDuplicateNames: Bool {
+        // Check for duplicates within new babies
+        let newBabyNames = babies.map { $0.baby.name.lowercased() }
+        if Set(newBabyNames).count != newBabyNames.count {
+            return true
+        }
+        
+        // Check against existing babies
+        let existingNames = Set(existingBabies.map { $0.name.lowercased() })
+        return !newBabyNames.filter { !$0.isEmpty }
+            .allSatisfy { !existingNames.contains($0) }
+    }
+
+    private func isDuplicateName(_ name: String, forBabyId id: Int) -> Bool {
+        let lowercaseName = name.lowercased()
+        
+        if existingBabies.contains(where: { $0.name.lowercased() == lowercaseName }) {
+            return true
+        }
+        
+        return babies.contains(where: { $0.id != id && $0.baby.name.lowercased() == lowercaseName })
+    }
+    
+    private func loadExistingBabies() async {
+        do {
+            existingBabies = try await standard.getBabies()
+        } catch {
+            errorMessage = "Failed to load existing babies: \(error.localizedDescription)"
+            showAlert = true
+        }
+        isLoading = false
     }
     
     private func saveBabies() async {
+        guard !hasDuplicateNames else {
+            errorMessage = "Each baby must have a unique name"
+            showAlert = true
+            return
+        }
+        
         do {
             try await standard.addBabies(babies: babies.map(\.baby))
             onboardingNavigationPath.nextStep()
