--- Feedbridge/Views/WeightEntryView.swift
@@ -1,10 +1,0 @@
-//
-//  WeightEntryView.swift
-//  Feedbridge
-//
-//  Created by Shamit Surana on 2/8/25.
-//
-// SPDX-FileCopyrightText: 2025 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
